Add files via upload
🚀 I'm excited to provide Zomato API Analysis, my most recent project! 🌮🍕🍜 

🎯 Objectives: 

This Zomato API Analysis is revolutionary for all those foodies who enjoy trying out different cuisines on a budget. It's meant to assist you in finding the most reasonably priced restaurants in the country for particular types of food. Explore in-depth information about the best places to eat in the nation, as well as information about the regions with the greatest density of eateries offering the food you love.

🤔 Key Questions Explored:

🗺️ Build a Country Map Table
🗓️ Build a Calendar Table with essential date-related columns
🌐 Numbers of Restaurants based on City and Country
📅 Restaurants opening trends based on Year, Quarter, and Month
⭐ Count of Restaurants based on Average Ratings
💰 Buckets based on Average Price and distribution of restaurants
🍽️ Percentage of Restaurants with "Table Booking"
🚚 Percentage of Restaurants with "Online Delivery"
📊 Develop charts based on Cuisines, City, and Ratings

🙌 Acknowledgments: 

I would like to thank my mentor Jalpa Desai for all of their help and assistance with this project! Thank you to Baroda Institute Of Technology - BIT for giving me the space and tools I needed to make this vision a reality. 

I look forward to hearing your opinions and suggestions! Together, let's continue to explore the world of flavors. 🌍🍽 
#FoodieAdventures #DataAnalysis #ZomatoAPI